Convert all config render functions into fn components
