--- src/projects/globalise/getAnnotationItem.ts
@@ -1,5 +1,0 @@
-import { AnnoRepoAnnotation } from "../../model/AnnoRepoAnnotation";
-
-export const getAnnotationItem = (annotation: AnnoRepoAnnotation) => {
-  return annotation.body.type;
-};
