+++ src/projects/mondriaan/AnnotationItemContent.tsx
@@ -7,7 +7,7 @@ interface GetAnnotationItemContentProps {
   annotation: AnnoRepoAnnotation;
 }
 
-export const GetAnnotationItemContent = (
+export const AnnotationItemContent = (
   props: GetAnnotationItemContentProps,
 ) => {
   function renderAnnotationItemContent() {
