+++ src/projects/mondriaan/Help.tsx
@@ -1,3 +1,3 @@
-export const GetHelp = () => {
+export const Help = () => {
   return <div>Placeholder</div>;
 };
