+++ src/projects/republic/AnnotationLinks.tsx
@@ -2,7 +2,7 @@ import { Link, useParams } from "react-router-dom";
 import { ScanBody } from "../../model/AnnoRepoAnnotation";
 import { useAnnotationStore } from "../../stores/annotation";
 
-export const GetAnnotationLinks = () => {
+export const AnnotationLinks = () => {
   const params = useParams();
 
   const annotations = useAnnotationStore((state) => state.annotations);
