--- src/projects/republic/RenderSearchInfoPage.tsx
@@ -1,3 +1,0 @@
-export const RenderSearchInfoPage = () => {
-  return <div>PLACEHOLDER</div>;
-};
