+++ src/projects/republic/SearchInfoPage.tsx
@@ -1,0 +1,3 @@
+export const SearchInfoPage = () => {
+  return <div>PLACEHOLDER</div>;
+};
